feat: moved message bubble related code to a separate component.
