+++ components/ChatBox.vue
@@ -10,35 +10,15 @@
         </div>
       </div>
 
-      <div class="chat-content z-0 px-4">
-        <div
-          v-for="message in messages"
-          :key="message.id"
-          class="chat"
-          :class="message.userId === 'user' ? 'chat-end' : 'chat-start'"
-        >
-          <div class="chat-image avatar">
-            <div class="w-10 rounded-full">
-              <img :src="userAvatar(message.userId)" />
-            </div>
-          </div>
-          <div class="chat-header mb-1">
-            <p class="inline mr-2" >{{ userName(message.userId) }}</p>
-            <time class="text-xs opacity-50">{{ formatDate(message.createdAt) }}</time>
-          </div>
-          <div
-            class="chat-bubble"
-            :class="message.userId === 'user' ? 'bg-gray-600' : 'bg-gray-900'"
-          >
-            <Markdown :source="message?.text" class="w-full" />
-          </div>
-        </div>
-      </div>
+      <chat-bubble
+        v-for="message in messages"
+        :key="message.id"
+        class="chat chat-content z-0 px-4"
+        :class="message.userId === 'user' ? 'chat-end' : 'chat-start'"
+        :message="message"
+      />
 
-      <div class="form-control sticky bottom-0 bg-gray-800 w-full px-4 pb-4 pt-2">
-          <label class="label" :class="usersTyping?.length ? '' : 'hide'">
-            <span class="label-text-alt">Botman is typing...</span>
-          </label>
+      <div class="form-control sticky bottom-0 bg-gray-800 w-full p-4 pt-8">
           <input
             @keydown.enter.prevent="emitMessage()"
             v-model="messageText"
@@ -52,14 +32,12 @@
 </template>
 
 <script setup lang="ts">
-
   import { ref, defineProps, defineEmits } from 'vue'
-  import Markdown from 'vue3-markdown-it'
   import type { Ref } from 'vue'
   import { Message, User } from "../types"
   import { nanoid } from "nanoid"
-  import TimeAgo from 'javascript-time-ago'
-  import de from 'javascript-time-ago/locale/en'
+  import ChatBubble from './ChatBubble.vue'
+
 
   const props = defineProps<{
     me: User,
@@ -84,27 +62,6 @@
     messageText.value = '';
   }
 
-  function userName(name: string): string {
-    if (name === 'user') {
-      return 'Me'
-    } else {
-      return 'Botman'
-    }
-  }
-  
-  function userAvatar(name: string): string {
-    if (name === 'user') {
-      return '/avatar.jpg'
-    } else {
-      return '/bot.jpg'
-    }
-  }
-
-  TimeAgo.addDefaultLocale(de)
-  const timeAgo = new TimeAgo('de-DE')
-  function formatDate(date: Date) {
-    return timeAgo.format(date)
-  }
   const messageBox = ref<HTMLElement>()
   watch(
     () => props.messages.length,
@@ -126,10 +83,6 @@
 
 .hide {
   visibility: hidden;
-}
-
-.scroll {
-  overflow: scroll;
 }
 
 </style>
